big refactoring in reader: now there are nodes for each special form
main reason for this being that the parser will need to perform a lot of operations on the AST that will need to be aware of what special forms we are working with. For example, I will need this to implement things related to scoping. Working with raw lists is unwieldy and working with Java classes from Java is as well, hence this middle ground that will help me keep things straightforward